requirements.txt
scrapy
psycopg2-binary